Move WPF-specific code to separate projects
The XPlot libraries had a reference to WPF-specific libraries. Unfortunately, this makes them
hard to use in cross-platform context - even though, HTML-based visualization is probably the
best way to get a nice cross-platform visualization for F#.

This commit moves the WPF-specific extensions for showing charts to separate DLLs (which could still
be included in the same NuGet package). This is the first step towards integrating XPlot with
FsLab (and hopefully making it standard cross-platform visualization library for F#!)